+++ app/Http/Controllers/KepalaToko/ProdukController.php
@@ -150,7 +150,16 @@ class ProdukController extends Controller
     {
         $item = Product::findOrFail($id);
 
+        if (
+            $item->relasiOrder()->exists()
+        ) {
+            toast('Data Produk yang memiliki riwayat transaksi tidak bisa dihapus.', 'error');
+            return redirect()->back();
+        }
+
         $item->delete();
+
+        toast('Data Produk berhasil dihapus.', 'success');
 
         return redirect()->route('item.index');
     }
